Update AuthService to API v2
